deploy and tests

+++ contracts/ChainlinkPriceOracle.sol
@@ -5,15 +5,16 @@ import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 contract ChainlinkPriceOracle {
     AggregatorV3Interface internal priceFeed;
 
-    constructor(){
-        priceFeed = AggregatorV3Interface(0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419);  //Price of ETH in USD      
+    constructor() {
+        priceFeed = AggregatorV3Interface(0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419); //Price of ETH in USD
     }
 
-    function getLatestPrice() public view returns(int) {
-        (   uint80 roundID,
-            int price,
-            uint startedAt,
-            uint timeStamp,
+    function getLatestPrice() public view returns (int256) {
+        (
+            uint80 roundID,
+            int256 price,
+            uint256 startedAt,
+            uint256 timeStamp,
             uint80 answeredInRound
         ) = priceFeed.latestRoundData();
         return price / 1e8;
